Combine Let and Subject modules into one
