+++ lib/rspec/core.rb
@@ -23,8 +23,7 @@ require_rspec['core/reporter']
 
 require_rspec['core/metadata_hash_builder']
 require_rspec['core/hooks']
-require_rspec['core/subject']
-require_rspec['core/let']
+require_rspec['core/memoized_helpers']
 require_rspec['core/metadata']
 require_rspec['core/pending']
 require_rspec['core/formatters']
